requirements.txt
flask
llama-cpp-python
openai
websockets
requests
psutil
python-dotenv
